+++ glue/utils/qt/colors.py
@@ -17,7 +17,7 @@ __all__ = ['mpl_to_qt4_color', 'qt4_to_mpl_color', 'cmap2pixmap',
            'QColormapCombo']
 
 
-def mpl_to_qt4_color(color, alpha=1.0):
+def mpl_to_qt4_color(color, alpha=None):
     """
     Convert a matplotlib color stirng into a Qt QColor object
 
@@ -37,9 +37,10 @@ def mpl_to_qt4_color(color, alpha=1.0):
         return QtGui.QColor(0, 0, 0, 0)
 
     cc = ColorConverter()
-    r, g, b = cc.to_rgb(color)
-    alpha = max(0, min(255, int(256 * alpha)))
-    return QtGui.QColor(r * 255, g * 255, b * 255, alpha)
+    r, g, b, a = cc.to_rgba(color)
+    if alpha is not None:
+        a = alpha
+    return QtGui.QColor(r * 255, g * 255, b * 255, a * 255)
 
 
 def qt4_to_mpl_color(qcolor):