+++ glue/viewers/table/qt/viewer_widget.py
@@ -4,25 +4,44 @@ import os
 import numpy as np
 
 from qtpy.QtCore import Qt
-from qtpy import QtGui, QtCore, QtWidgets
+from qtpy import QtCore, QtGui, QtWidgets
 from qtpy import PYQT5
-from glue.core.subset import ElementSubsetState
-from glue.core.edit_subset_mode import EditSubsetMode
+from matplotlib.colors import ColorConverter
+
+from glue.config import viewer_tool
+from glue.core.layer_artist import LayerArtistBase
 from glue.core import message as msg
+from glue.core import Data
+from glue.utils import nonpartial
 from glue.utils.qt import load_ui
 from glue.viewers.common.qt.data_viewer import DataViewer
 from glue.viewers.common.qt.toolbar import BasicToolbar
+from glue.viewers.common.qt.tool import CheckableTool
+from glue.core.subset import ElementSubsetState
+from glue.core.edit_subset_mode import EditSubsetMode
+from glue.core.state import lookup_class_with_patches
+from glue.utils.colors import alpha_blend_colors
+from glue.utils.qt import mpl_to_qt4_color
+from glue.core.exceptions import IncompatibleAttribute
+
+COLOR_CONVERTER = ColorConverter()
 
 
 class DataTableModel(QtCore.QAbstractTableModel):
 
-    def __init__(self, data):
+    def __init__(self, table_viewer):
         super(DataTableModel, self).__init__()
-        if data.ndim != 1:
-            raise ValueError("Can only use Table widget for one-dimensional data")
-        self._data = data
+        if table_viewer.data.ndim != 1:
+            raise ValueError("Can only use Table widget for 1D data")
+        self._table_viewer = table_viewer
+        self._data = table_viewer.data
         self.show_hidden = False
-        self.order = np.arange(data.shape[0])
+        self.order = np.arange(self._data.shape[0])
+
+    def data_changed(self):
+        top_left = self.index(0, 0)
+        bottom_right = self.index(self.columnCount(), self.rowCount())
+        self.dataChanged.emit(top_left, bottom_right)
 
     @property
     def columns(self):
@@ -53,6 +72,7 @@ class DataTableModel(QtCore.QAbstractTableModel):
             return None
 
         if role == Qt.DisplayRole:
+
             c = self.columns[index.column()]
             idx = self.order[index.row()]
             comp = self._data.get_component(c)
@@ -64,6 +84,29 @@ class DataTableModel(QtCore.QAbstractTableModel):
                 return comp[idx].decode('ascii')
             else:
                 return str(comp[idx])
+
+        elif role == Qt.BackgroundRole:
+
+            idx = self.order[index.row()]
+
+            # Find all subsets that this index is part of
+            colors = []
+            for layer_artist in self._table_viewer.layers[::-1]:
+                if layer_artist.visible:
+                    subset = layer_artist.layer
+                    try:
+                        if subset.to_mask(view=slice(idx, idx + 1))[0]:
+                            colors.append(subset.style.color)
+                    except IncompatibleAttribute as exc:
+                        layer_artist.disable_invalid_attributes(*exc.args)
+                    else:
+                        layer_artist.enabled = True
+
+            # Blend the colors using alpha blending
+            if len(colors) > 0:
+                color = alpha_blend_colors(colors, additional_alpha=0.5)
+                color = mpl_to_qt4_color(color)
+                return QtGui.QBrush(color)
 
     def sort(self, column, ascending):
         c = self.columns[column]
@@ -77,12 +120,46 @@ class DataTableModel(QtCore.QAbstractTableModel):
         self.layoutChanged.emit()
 
 
+class TableLayerArtist(LayerArtistBase):
+    def __init__(self, layer, table_viewer):
+        self._table_viewer = table_viewer
+        super(TableLayerArtist, self).__init__(layer)
+    def redraw(self):
+        self._table_viewer.model.data_changed()
+    def update(self):
+        pass
+    def clear(self):
+        pass
+
+
+@viewer_tool
+class RowSelectTool(CheckableTool):
+
+    tool_id = 'table:rowselect'
+    icon = 'glue_row_select'
+    action_text = 'Select row(s)'
+    tool_tip = ('Select rows by clicking on rows and pressing enter '
+                'once the selection is ready to be applied')
+    status_tip = 'CLICK to select, press ENTER to finalize selection'
+
+    def __init__(self, viewer):
+        super(RowSelectTool, self).__init__(viewer)
+        self.deactivate()
+
+    def activate(self):
+        self.viewer.ui.table.setSelectionMode(QtWidgets.QAbstractItemView.ExtendedSelection)
+
+    def deactivate(self):
+        self.viewer.ui.table.setSelectionMode(QtWidgets.QAbstractItemView.NoSelection)
+        self.viewer.ui.table.clearSelection()
+
+
 class TableWidget(DataViewer):
 
     LABEL = "Table Viewer"
 
     _toolbar_cls = BasicToolbar
-    tools = []
+    tools = ['table:rowselect']
 
     def __init__(self, session, parent=None, widget=None):
 
@@ -107,103 +184,94 @@ class TableWidget(DataViewer):
         else:
             hdr.setResizeMode(hdr.Interactive)
 
-        self.ui.table.clicked.connect(self._clicked)
-
         self.model = None
-        self.subset = None
-        self.selected_rows = []
+
+    def keyPressEvent(self, event):
+        if event.key() in [Qt.Key_Enter, Qt.Key_Return]:
+            self.finalize_selection()
+        super(TableWidget, self).keyPressEvent(event)
+
+    def finalize_selection(self):
+        model = self.ui.table.selectionModel()
+        selected_rows = [self.model.order[x.row()] for x in model.selectedRows()]
+        subset_state = ElementSubsetState(indices=selected_rows, data=self.data)
+        mode = EditSubsetMode()
+        mode.update(self._data, subset_state, focus_data=self.data)
+        self.ui.table.clearSelection()
 
     def register_to_hub(self, hub):
 
         super(TableWidget, self).register_to_hub(hub)
 
-        dfilter = lambda x: True
-        dcfilter = lambda x: True
-        subfilter = lambda x: True
+        def dfilter(x):
+            return x.sender.data is self.data
 
         hub.subscribe(self, msg.SubsetCreateMessage,
-                      handler=self._add_subset,
+                      handler=nonpartial(self._refresh),
                       filter=dfilter)
 
         hub.subscribe(self, msg.SubsetUpdateMessage,
-                      handler=self._update_subset,
-                      filter=subfilter)
+                      handler=nonpartial(self._refresh),
+                      filter=dfilter)
 
         hub.subscribe(self, msg.SubsetDeleteMessage,
-                      handler=self._remove_subset)
+                      handler=nonpartial(self._refresh),
+                      filter=dfilter)
 
         hub.subscribe(self, msg.DataUpdateMessage,
-                      handler=self.update_window_title)
+                      handler=nonpartial(self._refresh),
+                      filter=dfilter)
 
-    def _clicked(self, mode):
-        self._broadcast_selection()
+    def _refresh(self):
+        self._sync_layers()
+        self.model.data_changed()
 
-    def _broadcast_selection(self):
+    def _sync_layers(self):
 
-        if self.data is not None:
+        # For now we don't show the data in the list because it always has to
+        # be shown
 
-            model = self.ui.table.selectionModel()
-            self.selected_rows = [self.model.order[x.row()] for x in model.selectedRows()]
-            subset_state = ElementSubsetState(self.selected_rows)
+        for layer_artist in self.layers:
+            if layer_artist.layer not in self.data.subsets:
+                self._layer_artist_container.remove(layer_artist)
 
-            mode = EditSubsetMode()
-            mode.update(self.data, subset_state, focus_data=self.data)
-
-    def _add_subset(self, message):
-        self.subset = message.subset
-        self.selected_rows = self.subset.to_index_list()
-        self._update_selection()
-
-    def _update_selection(self):
-
-        self.ui.table.clearSelection()
-        selection_mode = self.ui.table.selectionMode()
-        self.ui.table.setSelectionMode(QtWidgets.QAbstractItemView.MultiSelection)
-
-        # The following is more efficient than just calling selectRow
-        model = self.ui.table.selectionModel()
-        for index in self.selected_rows:
-            index = self.model.order[index]
-            model_index = self.model.createIndex(index, 0)
-            model.select(model_index,
-                         QtCore.QItemSelectionModel.Select | QtCore.QItemSelectionModel.Rows)
-
-        self.ui.table.setSelectionMode(selection_mode)
-
-    def _update_subset(self, message):
-        self._add_subset(message)
-
-    def _remove_subset(self, message):
-        self.ui.table.clearSelection()
-
-    def _update_data(self, message):
-        self.set_data(message.data)
-
-    def unregister(self, hub):
-        pass
+        for subset in self.data.subsets:
+            if subset not in self._layer_artist_container:
+                self._layer_artist_container.append(TableLayerArtist(subset, self))
 
     def add_data(self, data):
         self.data = data
-        self.set_data(data)
+        self.setUpdatesEnabled(False)
+        self.model = DataTableModel(self)
+        self.ui.table.setModel(self.model)
+        self.setUpdatesEnabled(True)
+        self._sync_layers()
         return True
 
     def add_subset(self, subset):
         return True
 
-    def set_data(self, data):
-        self.setUpdatesEnabled(False)
-        self.model = DataTableModel(data)
-        self.model.layoutChanged.connect(self._update_selection)
-        self.ui.table.setModel(self.model)
-        self.setUpdatesEnabled(True)
+    def unregister(self, hub):
+        pass
 
     def closeEvent(self, event):
         """
-        On close, QT seems to scan through the entire model
+        On close, Qt seems to scan through the entire model
         if the data set is big. To sidestep that,
         we swap out with a tiny data set before closing
         """
-        from glue.core import Data
         d = Data(x=[0])
         self.ui.table.setModel(DataTableModel(d))
         event.accept()
+
+    def restore_layers(self, rec, context):
+        # For now this is a bit of a hack, we assume that all subsets saved
+        # for this viewer are from dataset, so we just get Data object
+        # then just sync the layers.
+        for layer in rec:
+            c = lookup_class_with_patches(layer.pop('_type'))
+            props = dict((k, context.object(v)) for k, v in layer.items())
+            layer = props['layer']
+            self.add_data(layer.data)
+            break
+        self._sync_layers()